Refined placing of terrain
Changed the way that it found a speciefic point in the terrain to be
a bit better. There is a problem when you modify a chunk boundry that it
only updates one of the two chunks so there ends up with a hole in the
world. The camera controls are also really bad and need to be fixed.